fix: rename trackable classes
