+++ src/Services/Trackable.php
@@ -4,20 +4,7 @@ declare(strict_types = 1);
 
 namespace Onyx\Services;
 
-use Ramsey\Uuid\Uuid;
-
-abstract class Trackable implements TrackableObject
+interface Trackable
 {
-    private
-        $trackingId;
-
-    public function __construct()
-    {
-        $this->trackingId = (Uuid::uuid4())->toString();
-    }
-
-    public function getTrackingId()
-    {
-        return $this->trackingId;
-    }
+    public function getTrackingId(): string;
 }
