--- src/Services/TrackableObject.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types = 1);
-
-namespace Onyx\Services;
-
-interface TrackableObject
-{
-    public function getTrackingId(): string;
-}
